Add. 🦄스웨거 세팅
