+++ backend/src/app.service.ts
@@ -1,8 +1,8 @@
-import {Injectable} from "@nestjs/common";
+import { Injectable } from '@nestjs/common';
 
 @Injectable()
 export class AppService {
   getHello(): string {
-    return "Hello World!";
+    return 'Hello World!';
   }
 }
