+++ backend/src/controller/mission/mission.module.ts
@@ -1,9 +1,12 @@
 import { Module } from '@nestjs/common';
 import { MissionService } from './mission.service';
 import { MissionController } from './mission.controller';
+import { TypeOrmModule } from '@nestjs/typeorm';
+import { Group, User } from '../../database/entities';
 
 @Module({
+  imports: [TypeOrmModule.forFeature([Group])],
   controllers: [MissionController],
-  providers: [MissionService]
+  providers: [MissionService],
 })
 export class MissionModule {}
